+++ projects/ngx-collection/src/lib/index.ts
@@ -1,7 +1,5 @@
 export { Collection } from './collection';
 export { ObjectsComparatorFn, ObjectsComparator, Comparator, DuplicateError } from './comparator';
 export { createEffect } from './create-effect';
-export { createEffect as effect } from './create-effect';
-export { createEffect as sideEffect } from './create-effect';
 export type * from './types';
 export { equalPrimitives } from './helpers';