+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### 5.0.0
+* Angular v20 is now supported;
+* The generator callback now has the second argument, `callbacks`. It contains an object with callbacks `success` and `error` - the observable you provide to `createEffect()` can use them to notify the effect caller;
+* The second argument of `createEffect()` got new fields:
+* * `onSuccess?: (v: unknown) => void` - will be called when `callbacks.success` is called;
+* * `onError?: (v: unknown) => void` - will be called when `callbacks.error` is called;
+* * `onFinalize?: () => void` - will be called when either `callbacks.success` or `callbacks.error` is called;
+
+#### BREAKING CHANGES
+* `createEffect.getEffectFor()` renamed to `createEffect.asObservable()`;
+* `createEffect.next$` is removed;
+* `createEffect.error$` is removed.
+
 ### 4.3.2
 Add Angular v19 to the list of supported versions in dependencies.